+++ es5-persistence/src/main/java/com/netflix/conductor/elasticsearch/es5/ElasticSearchV5Module.java
@@ -18,8 +18,8 @@ package com.netflix.conductor.elasticsearch.es5;
 import com.google.inject.AbstractModule;
 
 import com.netflix.conductor.dao.IndexDAO;
-import com.netflix.conductor.dao.es5.index.ElasticSearchDAOV5;
-import com.netflix.conductor.dao.es5.index.ElasticSearchRestDAOV5;
+import com.netflix.conductor.dao.ElasticSearchDAOV5;
+import com.netflix.conductor.dao.ElasticSearchRestDAOV5;
 import com.netflix.conductor.elasticsearch.ElasticSearchConfiguration;
 import com.netflix.conductor.elasticsearch.EmbeddedElasticSearchProvider;
