--- kafka-persistence/src/main/java/com/netflix/conductor/dao/kafka/index/utils/DocumentTypes.java
@@ -1,10 +1,0 @@
-package com.netflix.conductor.dao.kafka.index.utils;
-
-public class DocumentTypes {
-
-    public static final String WORKFLOW_DOC_TYPE = "workflow";
-    public static final String TASK_DOC_TYPE = "task";
-    public static final String LOG_DOC_TYPE = "task_log";
-    public static final String EVENT_DOC_TYPE = "event";
-    public static final String MSG_DOC_TYPE = "message";
-}
