--- kafka-persistence/src/main/java/com/netflix/conductor/dao/kafka/index/utils/OperationTypes.java
@@ -1,8 +1,0 @@
-package com.netflix.conductor.dao.kafka.index.utils;
-
-public class OperationTypes {
-    public static final String CREATE = "CREATE";
-    public static final String UPDATE = "UPDATE";
-    public static final String DELETE = "DELETE";
-
-}
