+++ kafka-persistence/src/main/java/com/netflix/conductor/kafka/KafkaModuleProvider.java
@@ -1,11 +1,12 @@
 package com.netflix.conductor.kafka;
 
 import com.google.inject.AbstractModule;
-import com.netflix.conductor.dao.ProducerDAO;
-import com.netflix.conductor.dao.kafka.index.KafkaModule;
-import com.netflix.conductor.dao.kafka.index.producer.KafkaProducer;
+import com.netflix.conductor.dao.IndexDAO;
+import com.netflix.conductor.dao.KafkaDAO;
 import com.netflix.conductor.elasticsearch.ElasticSearchConfiguration;
+import com.netflix.conductor.elasticsearch.EmbeddedElasticSearchProvider;
 import com.netflix.conductor.elasticsearch.SystemPropertiesElasticSearchConfiguration;
+import com.netflix.conductor.elasticsearch.es5.EmbeddedElasticSearchV5Provider;
 
 public class KafkaModuleProvider extends AbstractModule {
 
@@ -16,8 +17,8 @@ public class KafkaModuleProvider extends AbstractModule {
     protected void configure() {
         ElasticSearchConfiguration configuration = new SystemPropertiesElasticSearchConfiguration();
         if (configuration.getKafkaIndexEnable()) {
-            bind(ProducerDAO.class).to(KafkaProducer.class);
-            install(new KafkaModule(configuration));
+            bind(IndexDAO.class).to(KafkaDAO.class);
+            bind(EmbeddedElasticSearchProvider.class).to(EmbeddedElasticSearchV5Provider.class);
         }
     }
 }
