+++ kafka-persistence/src/main/java/com/netflix/conductor/kafka/index/utils/OperationTypes.java
@@ -1,0 +1,8 @@
+package com.netflix.conductor.kafka.index.utils;
+
+public class OperationTypes {
+    public static final String CREATE = "CREATE";
+    public static final String UPDATE = "UPDATE";
+    public static final String DELETE = "DELETE";
+
+}
